Added ContactViewController: -This view shows the contact fields, ready for registration: Telephone, Telephone Confirmation and Birthday.
Added the Contact Model:
-This model saves the customer contact information before performing a transaction with added warranties.
-Performs presence validations, and password confirmation.

Added WarrantyViewController:
-This view shows the available warranties for the scanned product.
- The view can register the selected warranty in the product list.

Added the Warranty Model:
-This model sets and retrieves the warranty info: cost, detail, percentage, sky and warrantyId.

Parse warranties in the FindItemModel when the scanned product has some.

Added a warranty flag to Session Model:
-This flag is for checking if a warranty is added to my product list or not.

Added warranties to the calculation of the payment amount in the Tools model:
- Calculates the amount of pay when warranties are added to the product list

Added a method for poping the warranties from the product list array in Tools:
-Deletes the warranties from this array when the array is sent to the WS to perform a sale.
(Warranties are set in the FindItemModel; they are not erased from the transaction)

Remove warranties from the data sent to promoRequest:
-Only the scanned products are sent to the promoRequest.

Disable the warranty flag from the ConsultSKUViewController:
-Send a false flag to the buscaProducto method, in order to receive the scanned product info without its available warranties.

Modified the cellForRowAtIndexPath in the ScanView:
- This was modified because the added warranties are now visible in my table view.

WarrantyViewController is presented after the product is scanned and has available warranties, in the Scan View.

ContactViewController is presented before the user sends the request for performing the Sale in the PaymentViewController.

Added a XML file for sending the customerData to the WS, when a transaction is made with more warranties.

Add support to the LiverPoolRequest Model, for processing transactions with more warranties:
- Handle the submit of the XML form for the customerData